Feature: Added support for self-signed certificates in the GrappaCommunicator
Added support for self-signed certificates in the GrappaCommunicator.
This allows the grader to connect to the middleware server even if the
certificate is self-signed.

Added a new configuration option to the proforma_settings.sh file:
ACCEPT_SELF_SIGNED_CERTS. This option can be set to YES or NO. If it is set
to YES, the grader will accept self-signed certificates. If it is set to NO,
the grader will reject self-signed certificates.